Introducing ArmCpuGen tool (work in progress).
The goal of this tool is to be able to automatically generate tables for decoding, disassembling and executing instructions with their various addressing modes.
This tool will also produce a smaller set of function calls that will need to be produced by hand in a separate file to deal with the actual logic of instructions and addressing modes.
In a first time, this tool will try to procedurally reconstruct a table with the same values than instruction_tabdef from DeSmuME using only the ARM instruction documentation on GBATEK. Once a match is confirmed, it will produce the real tables and functions that DSEmul would like to have.