Reworking ARM addressing mode and instruction table format.
Instruction table entries are now exported as calls to macro INSTRUCTION which must be implemented on the engine side before including the tables.
Instruction entries contain the name of the selected instruction (which can be easily remapped to the exported Insn enum and the type for the addressing mode which can be manipulated using traits.